Cleaned up code, additional checks and optimization

+++ classes/controller/manage.php
@@ -78,24 +78,34 @@ class Controller_Manage extends \Cms\Controller_Template
 	 */
 	public function action_delete($id = false)
 	{
-		$message = \Ticker\Model_Message::find($id);
+		$redirect_url = \Uri::create('/admin/ticker/manage');
 
-		if ($message) {
-			try
-			{
-				$message->delete();
-				\Session::set_flash("success", "Message deleted");
+		// Check if ID is valid
+		if ($id = (int)$id) {
+			// Fetch the ticker message
+			$message = \Ticker\Model_Message::find($id);
 
-				\Response::redirect(\Uri::create("/admin/ticker/manage"));
+			// Check if message is existent (not null)
+			if ($message) {
+				try {
+					$message->delete();
+					\Session::set_flash('success', 'Message deleted');
+				}
+				catch (\Exception $e) {
+					\Session::set_flash('error', 'Unable to delete message');
+					\Log::error($e->getMessage().PHP_EOL.$e->getTraceAsString(), __METHOD__);
+				}
 			}
-			catch(\Exception $e)
-			{
-				\Session::set_flash("Unable to delete message");
+			else {
+				\Session::set_flash('error', 'Message you were trying to delete does not exist');
 			}
 		}
+		else {
+			\Session::set_flash('error', 'Invalid ticker message ID');
+		}
 
-		\Session::set_flash('error', 'Message you were trying to delete does not exist');
-		\Response::redirect(\Uri::create("/admin/ticker/manage"));
+		// We don't want anyone to stay in delete page as it doesn't have any content
+		\Response::redirect($redirect_url);
 	}
 
 	public function action_set_default($id = false)
